Move ObjectBehavior into own file
